Move core functionality to GPUArraysCore.
