+++ lib/GPUArraysCore/Project.toml
@@ -1,0 +1,7 @@
+name = "GPUArraysCore"
+uuid = "46192b85-c4d5-4398-a991-12ede77f4527"
+authors = ["Tim Besard <tim.besard@gmail.com>"]
+version = "0.1.0"
+
+[deps]
+Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
