+++ src/host/abstractarray.jl
@@ -1,27 +1,5 @@
 # core definition of the AbstractGPUArray type
 
-export AbstractGPUArray
-
-"""
-    AbstractGPUArray{T, N} <: DenseArray{T, N}
-
-Supertype for `N`-dimensional GPU arrays (or array-like types) with elements of type `T`.
-Instances of this type are expected to live on the host, see [`AbstractDeviceArray`](@ref)
-for device-side objects.
-"""
-abstract type AbstractGPUArray{T, N} <: DenseArray{T, N} end
-
-const AbstractGPUVector{T} = AbstractGPUArray{T, 1}
-const AbstractGPUMatrix{T} = AbstractGPUArray{T, 2}
-const AbstractGPUVecOrMat{T} = Union{AbstractGPUArray{T, 1}, AbstractGPUArray{T, 2}}
-
-
-# convenience aliases for working with wrapped arrays
-
-const WrappedGPUArray{T,N} = WrappedArray{T,N,AbstractGPUArray,AbstractGPUArray{T,N}}
-
-const AnyGPUArray{T,N} = Union{AbstractGPUArray{T,N}, WrappedGPUArray{T,N}}
-
 
 # input/output
